refactor(desktop): expose croquis public api
