+++ apps/desktop/src/demo/UiDemoPage.tsx
@@ -49,7 +49,7 @@ import {
   ImportCompletedModal,
 } from '../features/library-workspace/import';
 import { FolderSearchSelect } from '@/entities/folder';
-import { CroquisStartModal } from '../features/croquis/ui/CroquisStartModal';
+import { CroquisStartModal } from '@/features/croquis';
 import {
   AnalyticsGraphWrapper,
   ContributionGraph,
