+++ apps/desktop/src/features/croquis/index.ts
@@ -1,1 +1,2 @@
 export { CroquisStartModal } from './ui/CroquisStartModal';
+export { CroquisWindow } from './ui/CroquisWindow';
