+++ apps/desktop/src/pages/croquis/index.tsx
@@ -1,4 +1,4 @@
-import { CroquisWindow } from '../../features/croquis/ui/CroquisWindow';
+import { CroquisWindow } from '@/features/croquis';
 
 export function CroquisPage() {
   return <CroquisWindow />;
